+++ lib/src/main/java/cyoap_main/unit/command/CreateCommand.java
@@ -7,7 +7,7 @@ public class CreateCommand extends AbstractCommand{
 	public double localy;
 	public ChoiceSet choiceSet;
 	
-	public CreateCommand(double x, double y, double local_x, double local_y) {
+	public CreateCommand(float x, float y, double local_x, double local_y) {
 		choiceSet = new ChoiceSet(x, y);
 		this.localx = local_x;
 		this.localy = local_y;
